Documentation comments for Common
