+++ LocalyticsXamarin/LocalyticsXamarin.Common/InAppEventArgs.cs
@@ -1,0 +1,22 @@
+﻿using System;
+
+namespace LocalyticsXamarin.Common
+{
+    /// <summary>
+    /// In app event arguments.
+    /// </summary>
+    public class InAppEventArgs : EventArgs { } // No Extra Args.
+    /// <summary>
+    /// In app did display event arguments.
+    /// </summary>
+    public class InAppDidDisplayEventArgs : EventArgs { } // No Extra Args.
+    /// <summary>
+    /// In app will dismiss event arguments.
+    /// </summary>
+    public class InAppWillDismissEventArgs : EventArgs { } // No Extra Args.
+    /// <summary>
+    /// In app did dismiss event arguments.
+    /// </summary>
+    public class InAppDidDismissEventArgs : EventArgs { } // No Extra Args.
+
+}
